requirements.txt
django 
djangorestframework
stripe
python-dotenv

